requirements.txt
dropbox==11.13.0
tweepy==4.0.0a0
requests==2.25.1
pytz==2021.1
loguru==0.5.3
dadjokes==1.3.2
discord.py==1.7.2
discord==1.7.3
Pillow==8.3.0
